+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.6] - 2021-07-15
+### Fixed
+- Could not unlink Certificate from Role when changed to a non Active or Grace Period status
+- handle CoPersonRole PUT action
+
 ## [0.3.5] - 2021-06-08
 ### Fixed
 - Correct VOMs URL preview form